\name{addLayer}
\alias{addLayer}
\title{OSMMap}
\usage{
  addLayer(OSMMap1, OSMMap2)
}
\arguments{
  \item{OSMMap1}{a map object resulting from the call to
  OSMMap}

  \item{OSMMap2}{another map object resulting from a call
  to OSMMap}
}
\description{
  Take two maps and mesh them so that the markers are on
  different layers
}
\details{
  OpenStreetMap Layer adding function
}
\author{
  Theo Van Rooy theo
}

