% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.JSON.OSMMap}
\alias{as.JSON.OSMMap}
\title{asJSON.OSMMap}
\usage{
asJSON.OSMMap(x,...)
}
\arguments{
\item{x}{is an OSMMap object resulting from the call to OSMMap or addLayers}
}
\description{
Return the geoJSON component of the OSMMap.
}
\details{
OpenStreetMap return json for plotting in opencpu or other.
}
\author{
Theo Van Rooy <theo@royaltyanalytics.com>
}

