% Generated by roxygen2 (4.0.2): do not edit by hand
\name{choroplethMap}
\alias{choroplethMap}
\title{choroplethMap}
\usage{
choroplethMap(df, state, plotVariable,...,county, color)
}
\arguments{
\item{state}{the name of the column in the df that provides the state name in full state name format (e.g. colorado)}

\item{plotVariable}{the name of the column in the df that provides the variable which can be used to produce heat map colored tiles}

\item{county}{the name of the column in the df that provides the county name in full county name format (e.g. weld)}

\item{color}{the name of the column which provides 6 digit (no alpha) hex colors for the state/county}

\item{palette}{'red', 'green', 'blue' determines the shading of the heat colors.  White = lowest, Full color = highest}

\item{layerName}{a character vector specifying the name of the layer}
}
\description{
Takes a df with a column marked for state and makes a choropleth plot for it with the color specified.
}
\details{
OpenStreetMap State and county level choropleth shape creating function
}
\examples{
\dontrun{
state  = choroplethMap(df, 'state')
counties = choroplethMap(df, 'state', 'county')
}
}
\author{
Theo Van Rooy <theo@royaltyanalytics.com>
}

