% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isServerRunning}
\alias{isServerRunning}
\title{isServerRunning}
\usage{
isServerRunning()
}
\description{
Check to the see if the R server is running.
}
\details{
OpenStreetMap Internal Function
}
\author{
Theo Van Rooy <theo@royaltyanalytics.com>
}

