\name{OSMMap.httpd.handler}
\alias{OSMMap.httpd.handler}
\title{isServerRunning}
\usage{
  OSMMap.httpd.handler(path, query)
}
\arguments{
  \item{path}{a path to run the webserver in}

  \item{query}{not used}

  \item{...}{additional arguments}
}
\description{
  Setup the R webserver properly.
}
\details{
  OpenStreetMap Internal Function for handling http
  requests
}
\author{
  Theo Van Rooy <theo
}

