\name{plot.OSMMap}
\alias{plot}
\alias{plot.OSMMap}
\title{plot.OSMMap}
\usage{
  plot(x,...)
}
\arguments{
  \item{x}{is an OSMMap object resulting from the call to
  OSMMap or addLayers}

  \item{...}{further plotting arguments}
}
\description{
  Plot the OSMMap Object on an HTML page and display it.
}
\details{
  OpenStreetMap Plotting for standalone usage.
}
\author{
  Theo Van Rooy <theo
}

