\name{print.OSMMap}
\alias{print}
\alias{print.OSMMap}
\title{print.OSMMap}
\usage{
  print(x,...)
}
\arguments{
  \item{x}{an OSMMap object resulting from the call to
  OSMMap or addLayers}

  \item{...}{further arguments}
}
\description{
  print the OSMMap Object to it's corresponding HTML/JS
  code.
}
\details{
  OpenStreetMap Printing (for use in RHTML or RMarkDown
  reports)
}
\author{
  Theo Van Rooy <theo
}

