% Generated by roxygen2 (4.0.2): do not edit by hand
\name{OSMMap}
\alias{OSMMap}
\title{OSMMap}
\usage{
OSMMap(df, ..., lat="lat", long="long", size=8, color="#00FFFF",popup='', group=NULL, layer="layer1")
}
\arguments{
\item{df}{is a dataFrame containing (at the least) two features called "lat" and "long"}

\item{...}{other parameters which can be passed into the mapping function see details.}

\item{lat}{indicates the column name of the df that gives the latitude of points or lines.}

\item{long}{indicates the column name of the df that gives the latitude of the points lines.}

\item{size}{indicates which column in the df gives the size of the marker, automatically scaled.}

\item{color}{indicates a column in the df which can be factored and split on for colors}

\item{popup}{indicates which column of the df provides text to place in the onClick popup for each map marker.}

\item{group}{indicates the column of the df that gives the line label for interpreting sequential points as a line rather than many markers.}

\item{layer}{name for the layer to appear on the OSMMap Plot.}

\item{groupType}{indicates whether the grouping is for distinct lines or a closed off and infilled polygon options (LineString, Polygon).}

\item{tileLayer}{specifies a tile set to use as the basemap, defaults to Open Street Maps.}

\item{colorByFactor}{default T, picks a rainbow color for each unique color specified, F means it uses the hex colors specified.}
}
\description{
The base level function for creating an interactive map with Open Street Maps
}
\details{
OpenStreetMap Mapping utility
}
\examples{
\dontrun{
#generate rand lats and longs near Denver,CO
df = data.frame(lat=runif(10, 38,40), long=-runif(10, 104,106),
               size=runif(10)*20, color=sample(rainbow(3), 10, replace=TRUE),
               group=sample(1:3, 10,replace=TRUE))
#base plot
plot(OSMMap(df))

#plot with specified size and color for markers
plot(OSMMap(df, size='size', color='color'))

#plot out a set of lines
linePlot = OSMMap(df, group='group', groupType="line")
polygonPlot = OSMMap(df, group='group', groupType='polygon')
plot(linePlot)
}
}
\author{
Theo Van Rooy theo@royaltyanalytics.com
}

